chore: extract components to `QuoteFormContent`
